FastAPI/requirements.txt
torch>=2.1.0
transformers>=4.39.0
torchvision
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.42.0
Pillow>=10.0.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.6


